Fix ORM schema type mismatches and add VIEW models
Phase 1: VIEW Models (CRITICAL)
- Created xras_views.py with 7 corrected VIEW models
  - XrasUserView, XrasRoleView, XrasActionView, XrasAllocationView
  - XrasHpcAllocationAmountView, XrasRequestView, CompActivityChargeView
- VIEW models now match actual database view schemas
- Deprecated old XRAS models that were incorrectly defined as tables
- Removed relationships from Allocation and User to old XRAS models

Phase 2: Type Mismatches (HIGH)
- Changed 33 columns from NUMERIC to Float (database uses FLOAT)
  - Allocations: amount, requested_amount, transaction_amount
  - All charge tables: terabyte_year, charge, core_hours
  - All charge summaries: charges, terabyte_years, core_hours
  - Machine/Queue: factor_value, wall_clock_hours_limit, time_limit_hours
- Changed 10 columns from DateTime to Date (summary tables)
  - All *_charge_summary tables: activity_date
  - All *_charge_summary_status tables: activity_date

Files modified: 16 ORM model files
Total changes: 41 type fixes + 7 new VIEW models

Tools added:
- tests/fix_schema_types.py - Automated type fix script
- tests/SCHEMA_FIX_PLAN.md - Comprehensive fix plan documentation

Note: VIEW models may need additional work to resolve import conflicts
with old models. This commit fixes the schema definitions; import
resolution will be completed in next commit.

+++ python/sam/activity/archive.py
@@ -69,8 +69,8 @@ class ArchiveCharge(Base):
     archive_activity_id = Column(Integer, ForeignKey('archive_activity.archive_activity_id'),
                                  nullable=False, unique=True)
     user_id = Column(Integer, ForeignKey('users.user_id'), nullable=False)
-    charge = Column(Numeric(22, 8))
-    terabyte_year = Column(Numeric(22, 8))
+    charge = Column(Float)
+    terabyte_year = Column(Float)
     activity_date = Column(DateTime)
 
     account = relationship('Account', back_populates='archive_charges')
